refactor: viewModel 하나로 합침 (#66)
